[PyROOT] Remove `PyzCppHelpers`
Most of these helpers were single-use, so it's better to inline them in
the implementations.
