--- bindings/pyroot/pythonizations/src/PyzCppHelpers.hxx
@@ -1,27 +1,0 @@
-// Author: Danilo Piparo CERN  08/2018
-
-/*************************************************************************
- * Copyright (C) 1995-2018, Rene Brun and Fons Rademakers.               *
- * All rights reserved.                                                  *
- *                                                                       *
- * For the licensing terms see $ROOTSYS/LICENSE.                         *
- * For the list of contributors see $ROOTSYS/README/CREDITS.             *
- *************************************************************************/
-
-#ifndef PYROOT_PYZCPPHELPERS
-#define PYROOT_PYZCPPHELPERS
-
-#include "CPyCppyy/API.h"
-
-#include "TClass.h"
-#include "ROOT/RConfig.hxx"
-
-#include <string>
-
-PyObject *CallPyObjMethod(PyObject *obj, const char *meth);
-PyObject *CallPyObjMethod(PyObject *obj, const char *meth, PyObject *arg1);
-PyObject *BoolNot(PyObject *value);
-TClass *GetTClass(const PyObject *pyobj);
-std::string GetScopedFinalNameFromPyObject(const PyObject *pyobj);
-
-#endif // PYROOT_PYZCPPHELPERS
